requirements.txt
python-telegram-bot==21.5
# for .env file support
python-dotenv==1.0.1